Check teams list not null

+++ devNest/src/main/java/es/urjc/dad/devNest/Internal_Services/GameJamService.java
@@ -285,14 +285,16 @@ public class GameJamService {
         if (gj != null) {
             List<TeamEntity> teams = gj.getTeams();
             List<Long> emptyTeams = new ArrayList<Long>();
-            if (!teams.isEmpty()) {
-                for (TeamEntity teamEntity : teams) {
-                    if (teamEntity.getMembers().size() == 0) {
-                        emptyTeams.add(teamEntity.getId());
+            if (teams != null) {
+                if (!teams.isEmpty()) {
+                    for (TeamEntity teamEntity : teams) {
+                        if (teamEntity.getMembers().size() == 0) {
+                            emptyTeams.add(teamEntity.getId());
+                        }
                     }
+                    teamRepository.deleteAllById(emptyTeams);
+                    gamejamRepository.save(gj);
                 }
-                teamRepository.deleteAllById(emptyTeams);
-                gamejamRepository.save(gj);
             }
         }
     }
